+++ ebustoolbox/models.py
@@ -11,6 +11,7 @@ from django.contrib.postgres.fields import ArrayField
 from django.db.models import QuerySet, Sum
 from django.dispatch import receiver
 from django.utils.timezone import make_aware
+from fast_update.query import FastUpdateManager
 
 MINIMAL_TRIP_DURATION_S = 60  # seconds
 
@@ -822,6 +823,7 @@ class Route(models.Model):
     geom = models.LineStringField(dim=3, srid=4326, null=True)
     distance = models.FloatField(default=None, null=False)
 
+    objects = FastUpdateManager()
     name = models.TextField(default=None, null=False, blank=True)
     name_short = models.TextField(default=None, null=True, blank=True)
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
@@ -918,6 +920,7 @@ class Trip(models.Model):
     class Meta:
         db_table = "Trip"
 
+    objects = FastUpdateManager()
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     route = models.ForeignKey(Route, null=False, on_delete=models.CASCADE)
 
@@ -999,7 +1002,7 @@ class StopTime(models.Model):
 
     """Intermediate stops of trips,
     which are not described by the arrival or departure of the trip"""
-
+    objects = FastUpdateManager()
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
 
     # When does the trip arrive at this station
@@ -1088,6 +1091,7 @@ class Event(models.Model):
     class Meta:
         db_table = "Event"
 
+    objects = FastUpdateManager()
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     vehicle_type = models.ForeignKey(VehicleType, null=False, on_delete=models.CASCADE)
     vehicle = models.ForeignKey(Vehicle, null=True, on_delete=models.CASCADE)