+++ ebustoolbox/tasks.py
@@ -1,6 +1,5 @@
 import collections
 import csv
-import json
 import shutil
 import traceback
 from argparse import Namespace
@@ -83,10 +82,55 @@ def input_files_to_database(cleaned_data: dict, request: HttpRequest):
     # Write the vehicle types to DB
     vehicles_to_db(simba_schedule.vehicle_types, django_scenario)
 
+    # ToDo Consistency check is nice, but should be handled in a more generic way.
+    # some cases are not handled like overlapping times, etc.
+    # Remove trips which have non unique times for arrival or departure per rotation
+    # Remove Rotations which dont start at the depot
+    filter_inconsistent_trips_and_rotations(simba_schedule)
+
     # Write the schedule including rotations and trips to the DB
     schedule_to_db(simba_schedule, django_scenario)
 
     return django_scenario, simba_schedule, original_args
+
+
+# ToDo Do somewhere else?
+def filter_inconsistent_trips_and_rotations(simba_schedule):
+    # Some filter functions to handle messy bvg input
+    counter = 0
+    del_rots = []
+    for key, rotation in simba_schedule.rotations.items():
+        depart_times = [t.departure_time for t in rotation.trips]
+        arrival_times = [t.arrival_time for t in rotation.trips]
+        start = 0
+        while True:
+            for i, _ in enumerate(rotation.trips[start:]):
+                i = i + start
+                if (
+                    depart_times.count(rotation.trips[i].departure_time) > 1
+                    or arrival_times.count(rotation.trips[i].arrival_time) > 1
+                ):
+                    break
+            else:
+                rotation.trips = list(sorted(rotation.trips, key=lambda x: x.departure_time))
+                break
+            counter += 1
+            rotation.trips.pop(i)
+            depart_times = [t.departure_time for t in rotation.trips]
+            arrival_times = [t.arrival_time for t in rotation.trips]
+            start = i
+
+        if (
+            rotation.trips[0].departure_name not in simba_schedule.stations
+            or rotation.trips[-1].arrival_name not in simba_schedule.stations
+        ):
+            del_rots.append(key)
+    print(
+        f"Deleting {len(del_rots)} rotations since they dont start or end at electrified station:{del_rots}"
+    )
+    for rot_id in del_rots:
+        del simba_schedule.rotations[rot_id]
+    print(f"{counter} trips deleted") if counter > 0 else print()
 
 
 def temperatures_to_db(
@@ -167,7 +211,27 @@ def get_schedule_from_db(django_scenario: Scenario) -> tuple[simba.schedule.Sche
     # Create soc dispatcher
     schedule.init_soc_dispatcher(args)
 
-    schedule.assign_only_new_vehicles()
+    # Database should contain assigned vehicles already
+    for rot in schedule.rotations.values():
+        assert rot.vehicle_id is not None
+
+    # This is done since vehicle counts were generated in vehicle_assignment.
+    # ToDo replace with simba call
+    # Calculate vehicle counts
+    # count number of vehicles per type
+    # used for unique vehicle id e.g. vehicletype_chargingtype_id
+    vehicle_type_counts = {
+        f"{vehicle_type}_{charging_type}": 0
+        for vehicle_type, charging_types in schedule.vehicle_types.items()
+        for charging_type in charging_types.keys()
+    }
+    unique_vids = {rot.vehicle_id for rot in schedule.rotations.values()}
+    for vid in unique_vids:
+        v_ls = vid.split("_")
+        vehicle_type_counts[f"{v_ls[0]}_{v_ls[1]}"] += 1
+    schedule.vehicle_type_counts = vehicle_type_counts
+
+    # schedule.assign_only_new_vehicles()
 
     return schedule, args
 
@@ -186,7 +250,12 @@ def get_rotations_and_trips_from_db(django_scenario, schedule, station_data) -> 
 
     for rot in Rotation.objects.filter(scenario=django_scenario):
         vehicle_type = rot.vehicle.vehicle_type.name_short
-        simba_rotation = SimbaRotation(id=rot.name, vehicle_type=vehicle_type, schedule=schedule)
+        simba_rotation = SimbaRotation(
+            id=rot.name,
+            vehicle_type=vehicle_type,
+            schedule=schedule,
+        )
+        simba_rotation.vehicle_id = rot.vehicle.name_short
         simba_rotation.charging_type = (
             EnumChargeType.OPPORTUNITY.value
             if rot.allow_opportunity_charging
@@ -662,20 +731,28 @@ def vary_depot_rotations(schedule) -> "collections.Iterable[SimbaRotation]":
     schedule.rotations = orig_rotations
 
 
+def run_toolchain_from_scenario(django_scenario: Scenario):
+    simba_schedule_db, args_db = get_schedule_from_db(django_scenario)
+    run_ebus_toolchain(simba_schedule_db, args_db, django_scenario.task_id)
+
+
+def run_simba_scenario(django_scenario: Scenario):
+    simba_schedule_db, args_db = get_schedule_from_db(django_scenario)
+    run_simba(simba_schedule_db, args_db, django_scenario.task_id)
+
+
 def _run_ebus_toolchain(schedule: SimbaSchedule, args, task_id):
     """Run the tool chain"""
-    # set report dir for first iteration
-    args.output_directory = Path(settings.UPLOAD_PATH) / task_id
-    report_dir = Path(settings.BASE_DIR, args.output_directory, "report_1")
     # call simba and eflips
-    run_simba(schedule, args, task_id, report_dir=report_dir)
-    # Currently some plots are generated and but not closed.
+
+    run_simba(schedule, args, task_id)
     # ToDo: Do this inside simba or spice ev
     plt.close()
 
     if settings.EFLIPS_USE:
 
         prev_events = [e.id for e in Event.objects.filter(scenario__task_id=task_id)]
+        print(f"Running eflips {datetime.now()}")
         run_eflips(task_id)
 
         eflips_assignment = get_assigned_vehicles(task_id, prev_events)
@@ -685,7 +762,7 @@ def _run_ebus_toolchain(schedule: SimbaSchedule, args, task_id):
         # TODO: currently report_directory is set in simba internally and is always report_1 for current purposes
         # (number changes by the amount of reports in the same fun of SimBA)
         # call simba with eflips results
-        run_simba(schedule, args, task_id, report_dir=report_dir)
+        run_simba(schedule, args, task_id)
 
 
 def get_assigned_vehicles(task_id: str, prev_events: List[Event]):
@@ -730,9 +807,9 @@ def get_assigned_vehicles(task_id: str, prev_events: List[Event]):
         if vehicle not in counted_vehicles:
             vt = vehicle.vehicle_type
             if vt.opportunity_charging_capable:
-                ct = EnumChargeType.OPPORTUNITY
+                ct = EnumChargeType.OPPORTUNITY.value
             else:
-                ct = EnumChargeType.DEPOT
+                ct = EnumChargeType.DEPOT.value
 
             vehicle_counter_dict[vt.name_short][ct] += 1
             counted_vehicles.add(vehicle)
@@ -755,52 +832,21 @@ def run_simba(
     schedule: SimbaSchedule,
     args,
     task_id,
-    report_dir=Path(".", "report"),
 ):
+    print(f"Running Simba {datetime.now()}")
     # TODO don't overwrite output on multiple function calls
+    args.output_directory = Path(settings.UPLOAD_PATH) / task_id
     args.attach_vehicle_soc = True
 
     db_scenario = Scenario.objects.get(task_id=task_id)
-
     scenario = schedule.run(args)
 
-    def dict_creator():
-        return dict(
-            departure_soc=None,
-            vehicle_type=[],
-            delta_soc=[],
-            arrival_soc=None,
-            minimal_soc=None,
-            charging_type=None,
-        )
-
-    # initialize eflips input
-    input_for_eflips = {
-        Rotation.objects.get(scenario=db_scenario, name=rot_id).id: dict_creator()
-        for rot_id in schedule.rotations
-    }
-
-    # Analyze schedules which are generated using different depot vehicles. I.e. every depot
-    # rotation is run with each vehicle to generate the consumption
-    for rot_id, rotation in schedule.rotations.items():
-        rotation.calculate_consumption()
-        db_rotation = Rotation.objects.get(scenario=db_scenario, name=rotation.id)
-        if rotation.charging_type == EnumChargeType.DEPOT:
-            input_for_eflips = depot_rotation_to_eflips_input(
-                db_rotation, db_scenario, input_for_eflips, rotation, schedule
-            )
-        else:
-            assert rotation.charging_type == EnumChargeType.OPPORTUNITY
-            input_for_eflips = opportunity_rotation_to_eflips_input(
-                db_rotation, db_scenario, input_for_eflips, rot_id, rotation, scenario, schedule
-            )
-
+    print(f"Plotting {datetime.now()}")
     schedule, scenario = simba.simulate.modes_simulation(schedule, scenario, args)
     db_scenario.finished = timezone.now()
     db_scenario.save()
-    # Create the file for eflips. This could be passed directly to eFlips by returning eflips_input
-    with open(Path(report_dir, "eflips_input.json"), "w") as f:
-        json.dump(input_for_eflips, f, indent=4)
+
+    print(f"Creating Simba Events {datetime.now()}")
 
     create_event_output(scenario, task_id)
 
@@ -852,6 +898,14 @@ def run_eflips(task_id) -> None:
     db_scenario = Scenario.objects.get(task_id=task_id)
 
     # Constructing the database URL manually
+    db_url = create_db_url()
+    generate_depot_layout(
+        db_scenario, database_url=db_url, charging_power=90, delete_existing_depot=False
+    )
+    simulate_scenario(db_scenario, database_url=db_url)
+
+
+def create_db_url():
     db_dict = settings.DATABASES["default"]
     engine = db_dict["ENGINE"].split(".")[-1]
     # sqlalchemy needs a translation of the engine
@@ -860,11 +914,7 @@ def run_eflips(task_id) -> None:
     db_url = (
         f"{engine}://{db_dict['USER']}:{db_dict['PASSWORD']}@{db_dict['HOST']}/{db_dict['NAME']}"
     )
-
-    generate_depot_layout(
-        db_scenario, database_url=db_url, charging_power=90, delete_existing_depot=False
-    )
-    simulate_scenario(db_scenario, database_url=db_url)
+    return db_url
 
 
 def get_timestep(simba_scenario: "SimbaScenario", timestamp: datetime) -> int:
@@ -901,6 +951,10 @@ def create_event_output(simba_scenario: "SimbaScenario", task_id):
     current_rotation = None
     for counter, vehicle_event in enumerate(sorted_vehicle_events):
         start_timestep = get_timestep(simba_scenario, vehicle_event.start_time)
+
+        event_time = vehicle_event.start_time
+        aware_start_time = make_aware(event_time) if not is_aware(event_time) else event_time
+
         try:
             if sorted_vehicle_events[counter + 1].vehicle_id == vehicle_event.vehicle_id:
                 end_time = sorted_vehicle_events[counter + 1].start_time
@@ -924,9 +978,8 @@ def create_event_output(simba_scenario: "SimbaScenario", task_id):
         if current_rotation is None:
             # first event must be a departure
             assert vehicle_event.event_type == "departure"
-            current_rotation = vehicle_trips.get(
-                departure_time=make_aware(vehicle_event.start_time)
-            ).rotation
+            current_rotation = vehicle_trips.get(departure_time=aware_start_time).rotation
+
             # Do not save events passed their rotation time. This is done by eflips
             last_arrival_time = (
                 Trip.objects.filter(rotation=current_rotation)
@@ -935,9 +988,10 @@ def create_event_output(simba_scenario: "SimbaScenario", task_id):
                 .arrival_time
             )
 
-        if make_aware(vehicle_event.start_time) == last_arrival_time:
+        if aware_start_time == last_arrival_time:
             # Set current rotation to none so new rotation will be looked up
             current_rotation = None
+            vehicle_trips = None
             continue
 
         end_timestep = min(get_timestep(simba_scenario, end_time), simba_scenario.step_i - 1)
@@ -953,16 +1007,14 @@ def create_event_output(simba_scenario: "SimbaScenario", task_id):
             )
 
         if vehicle_event.event_type == "arrival":
-            station = vehicle_trips.get(
-                arrival_time=make_aware(vehicle_event.start_time)
-            ).route.arrival_station
+            station = vehicle_trips.get(arrival_time=aware_start_time).route.arrival_station
 
             is_charging = vehicle_event.update["connected_charging_station"] is not None
             event_type = (
                 EventType.CHARGING_OPPORTUNITY if is_charging else EventType.STANDBY_DEPARTURE
             )
         elif vehicle_event.event_type == "departure":
-            trip = vehicle_trips.get(departure_time=make_aware(vehicle_event.start_time))
+            trip = vehicle_trips.get(departure_time=aware_start_time)
             event_type = EventType.DRIVING
         else:
             raise NotImplementedError("Unkown vehicle event type")